+++ src/files/file.h
@@ -4,9 +4,9 @@ constexpr int kMaxFilenameLength = 256;
 constexpr int kMaxPath = 260;
 
 SDL_RWops *openFile(const char *path, const char *mode = "rb");
-int loadFile(const char *path, void *buffer, int maxSize = -1, size_t skipBytes = 0, bool required = true);
-std::pair<char *, size_t> loadFile(const char *path, size_t bufferOffset = 0, size_t skipBytes = 0);
-bool saveFile(const char *path, void *buffer, size_t size);
+int loadFile(const char *path, void *buffer, int maxSize = -1, int skipBytes = 0, bool required = true);
+std::pair<char *, int> loadFile(const char *path, int bufferOffset = 0, int skipBytes = 0);
+bool saveFile(const char *path, void *buffer, int size);
 bool renameFile(const char *oldPath, const char *newPath);
 
 void setRootDir(const char *dir);
@@ -38,7 +38,7 @@ inline constexpr bool isFileSystemCaseSensitive()
 }
 
 int pathCompare(const char *path1, const char *path2);
-int pathNCompare(const char *path1, const char *path2, size_t count);
+int pathNCompare(const char *path1, const char *path2, int count);
 
 std::string joinPaths(const char *path1, const char *path2);
 bool fileExists(const char *path);
@@ -50,12 +50,12 @@ const char *getBasename(const char *path);
 
 struct FoundFile {
     std::string name;
-    size_t extensionOffset;
-    FoundFile(const std::string& name, size_t extensionOffset) : name(name), extensionOffset(extensionOffset) {}
+    int extensionOffset;
+    FoundFile(const std::string& name, int extensionOffset) : name(name), extensionOffset(extensionOffset) {}
 };
 using FoundFileList = std::vector<FoundFile>;
 
 FoundFileList findFiles(const char *extension, const char *dirName = nullptr,
-    const char **allowedExtensions = nullptr, size_t numAllowedExtensions = 0);
+    const char **allowedExtensions = nullptr, int numAllowedExtensions = 0);
 void traverseDirectory(const char *directory, const char *extension,
     std::function<bool(const char *, int, const char *)> f);
