+++ tests/src/mocks/mockFile.h
@@ -1,13 +1,13 @@
 #pragma once
 
 struct MockFile {
-    MockFile(const char *path, const char *data, size_t size) : path(path), data(data), size(size) {}
-    MockFile(const char *path, unsigned const char *data, size_t size)
+    MockFile(const char *path, const char *data, int size) : path(path), data(data), size(size) {}
+    MockFile(const char *path, unsigned const char *data, int size)
         : path(path), data(reinterpret_cast<const char *>(data)), size(size) {}
     MockFile(const char *path) : path(path) {}
     const char *path;
     const char *data = nullptr;
-    size_t size = 0;
+    int size = 0;
 };
 
 using MockFileList = std::vector<MockFile>;
@@ -18,8 +18,8 @@ void addFakeFiles(const MockFileList& files);
 void addFakeFile(const MockFile& file);
 void addFakeDirectory(const char *path);
 bool deleteFakeFile(const char *path);
-size_t getNumFakeFiles();
+int getNumFakeFiles();
 bool setFileAsCorrupted(const char *path, bool corrupted = true);
 bool fakeFilesEqualByContent(const char *path1, const char *path2);
-const char *getFakeFileData(const char *path, size_t& size, size_t& numWrites);
-size_t getFakeFileSize(const char *path);
+const char *getFakeFileData(const char *path, int& size, int& numWrites);
+int getFakeFileSize(const char *path);
